% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_quanta.R
\name{bootstrap_quanta}
\alias{bootstrap_quanta}
\title{Compute Boostrap Confidence Interval}
\usage{
bootstrap_quanta(
  set = NULL,
  sample_portion = 0.75,
  confint = 0.05,
  n_samples = 500,
  unround = TRUE,
  params = getOption("CONSTANTS_QUANTOGRAM")
)
}
\arguments{
\item{set}{a numeric vector of measurements from which quantum and bootstrap
should be computed}

\item{sample_portion}{The portion (<1) of the original data that should
be used for each sample}

\item{confint}{confidence level, 0.05 by default}

\item{n_samples}{number of bootstrap samples}

\item{unround}{TRUE/FALSE, declares if artificial noise should be added to
the measurements (T) or if they should be taken as is (F), defaults to TRUE.}

\item{params}{Constants used in the quantum calculation,
see getOption("CONSTANTS_QUANTOGRAM")}
}
\value{
a list containing the original quanta, a data.frame with the
bootstrap results and the confidence interval
}
\description{
This function employs the Monte Carlo bootstrap method to determine the
confidence interval that delimits the range of most possible quantums for
a given set of measurements.
}
\examples{
x <- rnorm(10) * 0.652
bootstrap_quanta(set = x, sample_portion = 0.5,
confint = 0.05, n_samples = 100)

}
