% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_error.R
\name{calculate_error}
\alias{calculate_error}
\title{Get sample error given a quantum}
\usage{
calculate_error(input, quantum)
}
\arguments{
\item{input}{numeric vector of measurement values}

\item{quantum}{numeric value of quantum}
}
\description{
Get sample error given a quantum
}
