% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantogram_orig.R
\name{calculate_quantogram_orig}
\alias{calculate_quantogram_orig}
\title{Calculate Kendall's cosine quantogram}
\usage{
calculate_quantogram_orig(x, params = getOption("CONSTANTS_QUANTOGRAM"))
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{params}{cosine quantogram parameters including "RNG_START", "RNG_END", "STEP", "Q_MIN"}
}
\description{
Calculate Kendall's cosine quantogram
}
