% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_selective_least_squares.R
\name{calculate_selective_least_squares}
\alias{calculate_selective_least_squares}
\title{Calculate selective least squares method (by Paakanen)}
\usage{
calculate_selective_least_squares(x, quanta)
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{quanta}{}
}
\description{

}
