% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sls.R
\name{calculate_sls}
\alias{calculate_sls}
\title{Calculate selective least squares method (by Paakanen)}
\usage{
calculate_sls(measurements, quanta)
}
\arguments{
\item{quanta}{a vector of quanta}

\item{x}{numeric vector of measurement values}
}
\value{

}
\description{
Calculate selective least squares method (by Paakanen)
}
