% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_boostrap.R
\name{compute_bootstrap}
\alias{compute_bootstrap}
\title{Compute boostrap confidence interval}
\usage{
compute_bootstrap(set, label, conf = 0.05, n_bootstrap_samples = 500,
  Q_MIN = Q_MIN, RNG_START = RNG_START, RNG_END = RNG_END,
  STEP = STEP)
}
\arguments{
\item{set}{measurements vector}

\item{label}{(optional) label of dataset}

\item{conf}{confidence level, 0.05 by default}

\item{n_bootstrap_samples}{number of bootstrap samples}
}
\description{
Compute boostrap confidence interval
}
