% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bootstrap_plot.R
\name{doc_bootstrap_plot}
\alias{doc_bootstrap_plot}
\title{Boostrap plot for documenting purpose}
\usage{
doc_bootstrap_plot(df = data_niche, sector_name, building_name = NULL,
  extra_suffix = "", n_bootstrap_samples = 500, conf = 0.05)
}
\arguments{
\item{conf}{}
}
\description{
Boostrap plot for documenting purpose
}
