% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erechtheion.R
\docType{data}
\name{Erechtheion}
\alias{Erechtheion}
\title{Erechtheion: Test Measurements}
\format{
A data frame with 19 rows and 5 variables:
\describe{
  \item{Name}{The object measured}
  \item{FeetAfterInscr}{the foot value the measurement should have
  according to the Inscription IG I^3 474 (see source)}
  \item{Measurement_m}{Actual measurement in meter}
  \item{Foot_m}{Length of foot that has been used for the measured object}
}
}
\source{
\url{J. Pakkanen, Deriving Ancient Foot Units from Building
Dimensions: a Statistical Approach Employing Cosine Quantogram
Analysis, 2002, in: G. Burenhult (Ed.), Archaeological Informatics:
Pushing the Envelope (Oxford 2002), 501--506 Table 1}
}
\usage{
Erechtheion
}
\description{
A dataset containing some measurements from the Eretheion in Athens.
This data was used in the publication referenced below and is a suitable
test dataset for quantatools.
}
\details{
Source: J. Pakkanen, Deriving Ancient Foot Units from Building Dimensions:
a Statistical Approach Employing Cosine Quantogram Analysis, 2002, in:
G. Burenhult (Ed.), Archaeological Informatics: Pushing the
Envelope (Oxford 2002), 501--506 Table 1
}
\keyword{datasets}
