% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_absolute_remainder.R
\name{least_absolute_remainder}
\alias{least_absolute_remainder}
\title{Title}
\usage{
least_absolute_remainder(z, q)
}
\arguments{
\item{z}{value 1}

\item{q}{value 2}
}
\value{
least absolute remainder between z and q
}
\description{
Title
}
