% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measerr_correction.R
\name{measerr_correction}
\alias{measerr_correction}
\title{Add a sequence of values around each measurement}
\usage{
measerr_correction(
  x,
  measurement_error = numeric(1),
  params = getOption("CONSTANTS_QUANTOGRAM")
)
}
\arguments{
\item{x}{a vector of measurements}

\item{measurement_error}{the possible error (same unit as measurements)}
}
\value{
x with additional sequences around each value
}
\description{
In cases where one is not confident that the measurements are consistent and
correct, this function adds a sequence of 'fake' measurements around each
value. Sequence runs between the values +- the value given as measurement
error.
}
