% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plots a Histogram of the bootstraped samples (within bootstrap_quanta)}
\usage{
plot_bootstrap(
  results,
  original_quanta,
  borders,
  binwidth = getOption("CONSTANTS_QUANTOGRAM")$STEP
)
}
\arguments{
\item{results}{result of the bootstrap sampling process within
`bootstrap_quanta()`}

\item{original_quanta}{original quanta as returned by `get_quantum()`}

\item{borders}{confidence interval}

\item{binwidth}{binwidth of the histogram (by default set to STEP used
in global package options)}
}
\value{

}
\description{
Plots a Histogram of the bootstraped samples (within bootstrap_quanta)
including lines for the confidence interval
}
