% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calculate_selective_least_squares.R
\name{plot_calculate_selective_least_squares}
\alias{plot_calculate_selective_least_squares}
\title{Plot and calculate SLS results}
\usage{
plot_calculate_selective_least_squares(x, sector_name,
  params = getOption("CONSTANTS_SLS"))
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{params}{}
}
\description{
Plot and calculate SLS results
}
