% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantogram.R
\name{plot_quantogram}
\alias{plot_quantogram}
\title{Plot cosine quantogram}
\usage{
plot_quantogram(df_results, sector_label)
}
\arguments{
\item{sector_label}{}
}
\description{
Plot cosine quantogram
}
