% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantatools-package.R
\docType{package}
\name{quantatools-package}
\alias{quantatools}
\alias{quantatools-package}
\title{quantatools: Analysis of quantum in archaeological data with cosine quantogram
    and related statistical methods}
\description{
The quantatools-package implements algorithms from cosine
    quantogram related research as proposed by D. G. Kendall in 1974. The
    aim of the package is to provide tools for researchers looking to find
    the base measurement unit (such as antique foot in building
    measurements) from a set of measurements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maciejkasinski/quantatools}
}

}
\author{
\strong{Maintainer}: Maciej Kasinski \email{maciejkasinski@gmail.com}

Other contributors:
\itemize{
  \item Lisa Steinmann \email{lisa.steinmann@rub.de} [contributor]
}

}
\keyword{internal}
