% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Get sample data from x for bootstrap}
\usage{
sample_bootstrap(x, decrease_by = 0)
}
\arguments{
\item{x}{vector, numeric. Original set of measurements}

\item{decrease_by}{integer. A value that should be subtracted from the
length of the original measurement vector (x)}
}
\value{
a sample of x decreased by `decrease_by`
}
\description{
Returns a subset of the data decreased by the absolute value given in
"decrease_by = ..."
}
\examples{
\dontrun{
x <- 1:100
sampleofx <- sample_bootstrap(x, decrease_by = 10)
length(sampleofx)}
}
\keyword{internal}
