% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unround_data.R
\name{unround_data}
\alias{unround_data}
\title{Unrounding Method}
\usage{
unround_data(input, depth = 0.01)
}
\arguments{
\item{input}{a numeric vector of values to be "unrounded"}

\item{depth}{the value by which the fictional "unrounded" measurement can
deviate}
}
\value{
input with more decimal numbers (randomized)
}
\description{
As proposed by Kendall (1974) data may be 'unrounded' before calculating
the quantogram by adding a random number between -0.005 and 0.005
to produce artificial noise.
}
\examples{
unround_data(c(1, 2, 3), depth = 0.01)
}
