% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantogram.R
\name{get_quantum}
\alias{get_quantum}
\title{Get a Quantum (single value) for a Set of Measurements}
\usage{
get_quantum(x, fofq = FALSE, unround = TRUE)
}
\arguments{
\item{x}{A numeric vector of measurement values for which
the quantum should be found}

\item{fofq}{TRUE / FALSE: Include f(q) in output or not}

\item{unround}{TRUE/FALSE, declares if artificial noise should be added to
the measurements (T) or if they should be taken as is (F), defaults to TRUE.}
}
\value{
The quantum of x
}
\description{
Get a Quantum (single value) for a Set of Measurements
}
\examples{
data <- 0.37 * rep(seq(3, 6, 2), 300)
get_quantum(data)
}
