% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot a Histogram of the Bootstrapped Samples}
\usage{
plot_bootstrap(
  results,
  title = "Bootstrapped Quanta",
  caption = NULL,
  binwidth = getOption("CONSTANTS_QUANTOGRAM")$STEP
)
}
\arguments{
\item{results}{result (list) of the bootstrap sampling process as returned by
`bootstrap_quanta()`, containing three lists: results = results of the
bootstrapped samples, quanta = original quanta, confidence interval = the
confidence interval.}

\item{title}{Title of the Plot}

\item{caption}{Caption of the Plot}

\item{binwidth}{binwidth of the histogram (by default set to STEP used
in global package options)}
}
\value{
A histogram with confidence intervals
}
\description{
Plot a Histogram of the bootstrapped samples
including lines for the confidence interval
}
