% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantogram.R
\name{plot_quantogram}
\alias{plot_quantogram}
\title{Plot cosine quantogram}
\usage{
plot_quantogram(results = NULL, title = "Quantogram")
}
\arguments{
\item{sector_label}{}
}
\value{

}
\description{
Plot cosine quantogram
}
\examples{
data("Erechtheion")
set <- Erechtheion$Measurement_m
results <- calculate_quantogram(set)
plot_quantogram(results, title = "This is a Quantogram")
}
