% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantogram.R
\name{plot_quantogram}
\alias{plot_quantogram}
\title{Plot Cosine Quantogram}
\usage{
plot_quantogram(results = NULL, title = "Quantogram", caption = NULL)
}
\arguments{
\item{results}{Resulting data.frame returned by calculate_quantogram(...)}

\item{title}{Title of the plot}

\item{caption}{Caption of the plot}
}
\value{
a plot object (ggplot2)
}
\description{
Plot Cosine Quantogram
}
\examples{
data("Erechtheion")
set <- Erechtheion$Measurement_m
results <- calculate_quantogram(set)
plot_quantogram(results, title = "This is a Quantogram")
}
