% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selective_least_squares.R
\name{selective_least_squares}
\alias{selective_least_squares}
\title{Plot and calculate SLS results}
\usage{
selective_least_squares(x, title = "...", params = getOption("CONSTANTS_SLS"))
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{title}{}
}
\value{

}
\description{
Plot and calculate SLS results
}
