% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unround_data.R
\name{unround_data}
\alias{unround_data}
\title{Unrounding method}
\usage{
unround_data(input, depth = 0.01)
}
\arguments{
\item{depth}{}
}
\value{
input with more decimal numbers (randomized)
}
\description{
Unrounding method
}
\examples{
unround_data(c(1, 2, 3), depth = 0.01)
}
